Create info
This code creates a Speedometer class that can be used to display the current speed, accelerate, and decelerate. You can set the maximum speed when creating an instance of the Speedometer class. Then you can call the accelerate and decelerate methods to change the speed, and display_speed will output the current speed.